requirements.txt
ultralytics>=8.0.0
flask>=2.0.0
opencv-python>=4.6.0
pillow>=8.0.0
numpy>=1.21.0
torch>=1.8.0
torchvision>=0.9.0
